+++ backend/src/main/java/io/centralweb/backend/service/ArticleService.java
@@ -3,13 +3,17 @@ package io.centralweb.backend.service;
 import io.centralweb.backend.dto.article.ArticleCreateDTO;
 import io.centralweb.backend.dto.article.ArticleDTO;
 import io.centralweb.backend.dto.article.ArticleUpdateDTO;
+import io.centralweb.backend.enums.UserRole;
 import io.centralweb.backend.events.ArticleCreateEvent;
 import io.centralweb.backend.exception.ObjectNotFoundException;
+import io.centralweb.backend.exception.ProfileIsNotTheOwnerException;
 import io.centralweb.backend.mapper.ArticleMapper;
 import io.centralweb.backend.model.Article;
 import io.centralweb.backend.model.Profile;
+import io.centralweb.backend.model.User;
 import io.centralweb.backend.repository.ArticleRepository;
 import io.centralweb.backend.repository.ProfileRepository;
+import io.centralweb.backend.repository.UserRepository;
 import jakarta.transaction.Transactional;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
@@ -23,13 +27,15 @@ import java.util.stream.Collectors;
 public class ArticleService {
     private final ArticleRepository articleRepository;
     private final ArticleMapper articleMapper;
+    private final UserRepository userRepository;
     private final ProfileRepository profileRepository;
     private final TagService tagService;
     private final ApplicationEventPublisher publisher;
 
-    public ArticleService(ArticleRepository articleRepository, ArticleMapper articleMapper, ProfileRepository profileRepository, TagService tagService, ApplicationEventPublisher publisher) {
+    public ArticleService(ArticleRepository articleRepository, ArticleMapper articleMapper, UserRepository userRepository, ProfileRepository profileRepository, TagService tagService, ApplicationEventPublisher publisher) {
         this.articleRepository = articleRepository;
         this.articleMapper = articleMapper;
+        this.userRepository = userRepository;
         this.profileRepository = profileRepository;
         this.tagService = tagService;
         this.publisher = publisher;
@@ -98,7 +104,7 @@ public class ArticleService {
                 .orElseThrow(() -> new ObjectNotFoundException("Artigo não encontrado"));
 
         if(!article.getProfile().getUser().getUserId().equals(userProfileId)) {
-            throw new RuntimeException("Você não tem permissão para isso");
+            throw new ProfileIsNotTheOwnerException("Você não tem permissão para isso");
         }
 
         articleMapper.updateArticleFromDTO(articleUpdated, article);
@@ -124,9 +130,16 @@ public class ArticleService {
         articleRepository.save(article);
     }
 
-    public void deleteArticleById(UUID articleId) {
+    public void deleteArticleById(UUID articleId, UUID userProfileId) {
+        User user = userRepository.findById(userProfileId)
+                .orElseThrow(() -> new ObjectNotFoundException("Usuário não existe"));
         Article article = articleRepository.findById(articleId)
                 .orElseThrow(() -> new ObjectNotFoundException("Artigo não encontrado"));
+
+        if(!article.getProfile().getUser().getUserId().equals(userProfileId) ||
+                !user.getRole().equals(UserRole.ADMIN)){
+            throw new ProfileIsNotTheOwnerException("Você não tem permissão para isso");
+        }
 
         articleRepository.delete(article);
     }
